+++ packages/orq-rc/RELEASES.md
@@ -10639,3 +10639,13 @@ Based on:
 - [typescript v4.12.0-rc.8] packages/orq-rc
 ### Releases
 - [NPM v4.12.0-rc.8] https://www.npmjs.com/package/@orq-ai/node/v/4.12.0-rc.8 - packages/orq-rc
+
+## 2026-07-08 21:22:52
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.789.0 (2.916.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.12.0-rc.9] packages/orq-rc
+### Releases
+- [NPM v4.12.0-rc.9] https://www.npmjs.com/package/@orq-ai/node/v/4.12.0-rc.9 - packages/orq-rc